A little pretentious
a mean.... if i wanna talk of multimedia, hypermedia and metamedia, i need to get involve on this shananigan, i stil dont know how write that word, f*ck it, this layer its a safe space. Lika an underworld, you see? on the surface the topics are the ships, even floating cities on the see, and here we are not in submarine field, even deeper, the abyssal, its more fun thinking on that way doesnt? open channels, get people smarter with the use of thecnology, make a better experience , given it in a sort of sauce or spicy,  a bit of storytelling, the web must be visualize as a playground with substante, our brain need textura, substance to remember the things, not just screens, needs gesture, movemment, colours, but displays, fields, inmersivenes, and the things are just on the edge, on the  verg? on the edge? ares just to explode in a whole new internet, the 3.0 its coming, the next chapter of the web, the next change of paradigm, and was a surprice that gain so much terrain on the race, a federate web woul be more ... utopic i guess, but technology reach the doors of technological singularity on the hand of AI, and the hardware its right on the corner, a full experiences, even holografic stereoscopi, its coming, i know it, i dont have the time and resources to do it, but if i have enough time, its my prediction its the enough of ... anticipate.... to give me time to develope it, there i go. 
Well, chapter one, chapter two its about metamedia and chapter three about my pretencious theory of semantic and cognitive fields, a theory thas its an expansion of wepir-worph linguistic relativism theory, where a system of symbols (beyond classic linguistic, like glyphs of graphic design such as emojis, but also UI UX elements, like a "like" button, stars, views, experience metrics, user - algorithm symbiosis, thatpart of the story, the system of symbols we create for communication throw all differents channels of human interaction, that... generate a mirror or conditional cognitive definition over humans... our way of thinking, someone in certain culture, certaing context, certain shows of television, touchable clickeable things, will have a view of the world different to a digital native, generational breach are good examples, someone very gammificated can structure the world in a different way and have a sort of dopamine regulation i think, i need to study it. 
So, system of symbols* , have a direct cognitive influence on the structure of the mind. And in time of a quickly changing on  UX, two new dimensions, and infinity technology, we will go slow on the race, a really really new internet its comming, and go through, trow? go ahead on this stream will be wild. 

(*) [ and metamedia on the deep] 

You see Arrival? good movie. 

and i branch, a lot. 🍄Rizoma vibes~